correzzioni varie e nuovo report

+++ sql/functions.sql
@@ -2,11 +2,11 @@ DROP FUNCTION costo;
 DROP FUNCTION ricavo;
 DROP FUNCTION margine;
 
-CREATE FUNCTION costo(costoGiornaliero INT, oreLavorate INT)
+CREATE FUNCTION costo(costoGiornaliero FLOAT, oreLavorate INT)
 RETURNS FLOAT DETERMINISTIC
 RETURN oreLavorate * costoGiornaliero / 8;
 
-CREATE FUNCTION ricavo(tariffaGiornaliera INT, oreLavorate INT)
+CREATE FUNCTION ricavo(tariffaGiornaliera FLOAT, oreLavorate INT)
 RETURNS FLOAT DETERMINISTIC
 RETURN oreLavorate * tariffaGiornaliera / 8;
